Inline a few private helpers
