+++ ui/dashboards/src/context/DashboardProvider/index.ts
@@ -15,3 +15,4 @@ export * from './dashboard-provider-api';
 export * from './DashboardProvider';
 export type { PanelGroupId, PanelGroupDefinition, PanelGroupItemId as LayoutItem } from './panel-group-slice';
 export type { PanelGroupEditor, PanelGroupEditorValues } from './panel-group-editor-slice';
+export type { DeletePanelDialogState } from './panel-delete-slice';
